webmk1/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
python-multipart==0.0.20
pdfplumber==0.11.4
google-api-python-client==2.160.0
google-auth==2.38.0
